bugfix: split the deploy workflow

+++ packages/contracts-bedrock/mantle-l1-upgrade.sh
@@ -1,4 +1,46 @@
 #!/bin/bash
 
-rm -rf deployments/devnetL1-mantle
-cp -r deployments/devnetL1-mantle-bak deployments/devnetL1-mantle
+a=$1
+
+if [ $a -eq 0 ]
+then
+  rm -rf deployments/devnetL1-mantle
+  cp -r deployments/devnetL1-mantle-bak deployments/devnetL1-mantle
+fi
+
+
+if [ $a -eq 1 ]
+then
+  echo $a
+  mv deploy deploy-deployment
+  mv deploy-upgrade-1 deploy
+  yarn hardhat --network devnetL1-mantle
+fi
+
+
+if [ $a -eq 2 ]
+then
+  echo $a
+  mv deploy deploy-upgrade-1
+  mv deploy-upgrade-2 deploy
+  yarn hardhat --network devnetL1-mantle
+fi
+
+
+if [ $a -eq 3 ]
+then
+  echo $a
+  mv deploy deploy-upgrade-2
+  mv deploy-upgrade-3 deploy
+  yarn hardhat --network devnetL1-mantle
+fi
+
+
+
+if [ $a -eq 4 ]
+then
+  echo $a
+  mv deploy deploy-upgrade-3
+  mv deploy deploy-upgrade-3
+  yarn hardhat --network devnetL1-mantle
+fi
